+++ client/pages/client/lesson.js
@@ -1,49 +1,50 @@
 import styles from "../../styles/Home.module.css";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import Sidebar from '../component/sidebar';
 
 export default function Lesson() {
-  const router = useRouter();
-
   return (
     <main>
       <div className={styles.containerCol}>
-        <div className={styles.column1}>
-            <Link href="/client/lesson" className={styles.button2}>
-              <div className={styles.japper}>Japper</div>
-            </Link>
-            
-            <Link href="/client/lesson" className={router.pathname === '/client/lesson' ? styles.activeButton1 : styles.button1}>
-              <div className={styles.lesson}>Lesson</div>
-            </Link>
-            <Link href="/client/game" className={router.pathname === '/client/game' ? styles.activeButton1 : styles.button1}>
-              <div className={styles.lesson}>Game</div>
-            </Link>
-            <Link href="/client/flashcard" className={router.pathname === '/client/flashcard' ? styles.activeButton1 : styles.button1}>
-              <div className={styles.lesson}>Flashcard</div>
-            </Link>
-            <Link href="/client/account" className={router.pathname === '/client/account' ? styles.activeButton1 : styles.button1}>
-              <div className={styles.lesson}>Account</div>
-            </Link>
-        </div>
+        <Sidebar/>
 
         <div className={styles.column2}>
           <div className={styles.box1}>
             <div className={styles.box2}>
-              <div className={styles.box3}><p className={styles.lesson}>Lesson</p></div>
-              <div className={styles.box4}><p className={styles.lesson}>content</p></div>
+              <div className={styles.box3}>Lesson fffffffff fffff</div>
+              <div className={styles.box4}>content kkkkkkkkkkkkkkkkkk lkkkkkkkkkkkkkkkk</div>
             </div>
 
-            <Link href="/client/lesson/vocab" className={styles.button3}>
-              <div className={styles.lesson}>Vocabulary</div>
-            </Link>
-            <Link href="/client/lesson/kanji" className={styles.button3}>
-              <div className={styles.lesson}>Kanji</div>
-            </Link>
-            <Link href="/client/lesson/grammar" className={styles.button3}>
-              <div className={styles.lesson}>Grammar</div>
-            </Link>
+            <Link href="/client/lesson/vocab" className={styles.button3}>Vocabulary</Link>
+            <Link href="/client/lesson/kanji" className={styles.button3}>Kanji</Link>
+            <Link href="/client/lesson/grammar" className={styles.button3}>Grammar</Link>
+            <Link href="/client/lesson/question" className={styles.button3}>Question</Link>
           </div>
+
+          <div className={styles.box1}>
+            <div className={styles.box2}>
+              <div className={styles.box3}>Lesson ffff fffff</div>
+              <div className={styles.box4}>content kkkkkkkkkkkkkkkkkk lkkkkkkkkkkkkkkkk</div>
+            </div>
+
+            <Link href="/client/lesson/vocab" className={styles.button3}>Vocabulary</Link>
+            <Link href="/client/lesson/kanji" className={styles.button3}>Kanji</Link>
+            <Link href="/client/lesson/grammar" className={styles.button3}>Grammar</Link>
+            <Link href="/client/lesson/question" className={styles.button3}>Question</Link>
+          </div>
+
+          <div className={styles.box1}>
+            <div className={styles.box2}>
+              <div className={styles.box3}>Lesson fffffffff fffff</div>
+              <div className={styles.box4}>content kkkkkkkkkkkkkkkkkk lkkkkkkkkkkkkkkkk</div>
+            </div>
+
+            <Link href="/client/lesson/vocab" className={styles.button3}>Vocabulary</Link>
+            <Link href="/client/lesson/kanji" className={styles.button3}>Kanji</Link>
+            <Link href="/client/lesson/grammar" className={styles.button3}>Grammar</Link>
+            <Link href="/client/lesson/question" className={styles.button3}>Question</Link>
+          </div>
+
         </div>
       </div>
     </main>
