+++ client/pages/client/lesson/kanji.js
@@ -1,11 +1,9 @@
 import styles from "../../../styles/Home.module.css";
 import modules from "../../../styles/Lesson.module.css";
-import Link from "next/link";
-import { useRouter } from "next/router";
 import {useState} from "react";
+import Sidebar from "../../component/sidebar";
 
 export default function Kanji() {
-  const router = useRouter();
 
   const [cards, setCards] = useState([
     { question: "Question 1", answer: "Answer 1" },
@@ -31,52 +29,7 @@ export default function Kanji() {
   return (
     <main>
       <div className={styles.containerCol}>
-        <div className={styles.column1}>
-          <Link href="/client/lesson" className={styles.button2}>
-            <div className={styles.japper}>Japper</div>
-          </Link>
-
-          <Link
-            href="/client/lesson"
-            className={
-              router.pathname === "/client/lesson/kanji"
-                ? styles.activeButton1
-                : styles.button1
-            }
-          >
-            <div className={styles.lesson}>Lesson</div>
-          </Link>
-          <Link
-            href="/client/game"
-            className={
-              router.pathname === "/client/game"
-                ? styles.activeButton1
-                : styles.button1
-            }
-          >
-            <div className={styles.lesson}>Game</div>
-          </Link>
-          <Link
-            href="/client/flashcard"
-            className={
-              router.pathname === "/client/flashcard"
-                ? styles.activeButton1
-                : styles.button1
-            }
-          >
-            <div className={styles.lesson}>Flashcard</div>
-          </Link>
-          <Link
-            href="/client/account"
-            className={
-              router.pathname === "/client/account"
-                ? styles.activeButton1
-                : styles.button1
-            }
-          >
-            <div className={styles.lesson}>Account</div>
-          </Link>
-        </div>
+        <Sidebar/>
 
         <div className={styles.column2}>
           <div className={modules.box3}>Kanji</div>
