+++ client/pages/index.js
@@ -1,33 +1,9 @@
-import styles from "../styles/Home.module.css";
-import Link from "next/link";
+import Lessons from '../pages/client/lesson';
 
 export default function Home() {
   return (
     <main>
-      <div className={styles.containerCol}>
-        <div className={styles.column1}>
-            <Link href="/" className={styles.button2}>
-              <div className={styles.japper}>Japper</div>
-            </Link>
-
-            <Link href="/client/lesson" className={styles.button1} >
-              <div className={styles.lesson}>Lesson</div>
-            </Link>
-            <Link href="/client/game" className={styles.button1}>
-              <div className={styles.lesson}>Game</div>
-            </Link>
-            <Link href="/client/flashcard" className={styles.button1}>
-              <div className={styles.lesson}>Flashcard</div>
-            </Link>
-            <Link href="/client/account" className={styles.button1}>
-              <div className={styles.lesson}>Account</div>
-            </Link>
-        </div>
-
-        <div className={styles.column2}>
-          column2
-        </div>
-      </div>
+      <Lessons/>
     </main>
   );
 }
