+++ client/styles/Home.module.css
@@ -7,174 +7,112 @@
   justify-content: center;
 }
 
-/*slideBar*/
-.column1 {
-  background-color: #ffffff;
-  border-right: 1px inset #b7b7b7;
-  height: 740px;
-  text-align: center;
-  justify-content: center;
-  box-sizing: border-box;
-}
-
 /*content*/
 .column2 {
   background-color: #ffffff;
-  height: auto;
+  height: 740px;
   text-align: center;
   justify-content: center;
   box-sizing: border-box;
   overflow-y: scroll; /*cho phép box bên trái di chuyển theo thanh kéo trang khi nội dung quá dài.*/
 }
 
-/*button cua lesson,game,flashcard,account*/
-.button1 {
-  text-decoration: none; /*bỏ các kiểu như underline*/
-  display: inline-block;
-  background: #fff4cf;
-  mix-blend-mode: normal;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 200px;
-  text-align: center;
-  justify-content: center;
-  margin: auto;
-  margin-bottom: 15px;
-}
-/* Định dạng phong cách cho khi nút được di chuột qua */
-.button1:hover {
-  background: #f6b51f;
-}
-/* Định dạng phong cách cho khi nút được chọn */
-.button1:focus {
-  background: #f6b51f;
-}
-.activeButton1 {
-  text-decoration: none; /*bỏ các kiểu như underline*/
-  display: inline-block;
-  background: #f6b51f;
-  mix-blend-mode: normal;
-  border: 1px solid #000000;
-  border-radius: 20px;
-  width: 200px;
-  text-align: center;
-  justify-content: center;
-  margin: auto;
-  margin-bottom: 15px;
-}
-
-/*button cua japper*/
-.button2 {
-  text-decoration: none; /*bỏ các kiểu như underline*/
-  display: inline-block;
-  mix-blend-mode: normal;
-  width: 200px;
-  text-align: center;
-  justify-content: center;
-  font-family: "Inter";
-  /*thu tu cac dong co anh huong den cach bai tri*/
-  margin: auto;
-  margin-top: 30px;
-  margin-bottom: 100px;
-}
-
-/*button cua kanji, grammar, vocab*/
+/*button cua kanji, grammar, vocab, question*/
 .button3 {
   text-decoration: none; /*bỏ các kiểu như underline*/
-  display: inline-block;
+  display: inline-flex;
   background: #ffffff;
   mix-blend-mode: normal;
   border: 1px solid #000000;
   border-radius: 20px;
   width: 200px;
+  height:50px;
   text-align: center;
   justify-content: center;
-  margin-right: 80px;
-  margin-left: 80px;
+  align-items: center;
+  margin-right: 30px;
+  margin-left: 30px;
+
+  font-family: 'MyCustomFont1';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 30px;
+  line-height: 36px;
+  color: #323030;
 }
 .button3:hover {
-  background-color: #d5d5d5;
+  background-color: #ffcb42;
 }
 
 /*box cua lesson*/
 .box1 {
-  display: flexbox;
+  display: inline-flexbox;
   width: 1140px;
-  height: 180px;
+  height: 290px;
   box-sizing: border-box;
-  border-bottom: 1px solid #615e5e;
+  border-top: 1px solid #878686;
+  border-bottom: 1px solid #878686;
   background: #ffffff;
   text-align: center;
+  align-items: center;
   justify-content: center;
   margin: auto;
 }
 
 .box2 {
-  display: flex;
+  display: inline-flex;
   box-sizing: border-box;
   background: #fff4cf;
   mix-blend-mode: normal;
   border: 1px solid #000000;
   border-radius: 20px;
-  width: 1100px;
-  height: 100px;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  margin-left: 20px;
+  width: 1050px;
+  height: 140px;
+  margin:40px auto 25px;
 }
 
 /*box cua so bai hoc*/
 .box3 {
-  display: inline-block;
-  width: 90px;
-  height: 36px;
-
+  width: 300px;
+  height: 40px;
+  margin-top: 20px;
+  margin-left: 20px;
+  margin-right: -60px;
+  
+  font-family: 'MyCustomFont1';
   text-align: center;
-  justify-content: center;
-  margin-left: 15px;
-  margin-top: -20px;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 32px;
+  display: inline-flex;
+  align-items: center;
+  color: #323030;
 }
 
 /*box cua ten bai hoc*/
 .box4 {
-  display: inline-block;
-  width: 180px;
-  height: 36px;
+  width: 700px;
+  height: 70px;
+  margin-top: 60px;
+  margin-left: -60px;
 
+  font-family: 'MyCustomFont1';
   text-align: center;
-  justify-content: center;
-  margin-left: 357px;
-  margin-top: 0px;
-}
-
-/*style cua chu*/
-.japper {
-  font-family: "Inter";
   font-style: normal;
-  font-weight: 900;
-  font-size: 50px;
-  line-height: 61px;
-  display: flex;
+  font-weight: 600;
+  font-size: 32px;
+  display: inline-flex;
   align-items: center;
-  text-align: center;
-  color: rgba(0, 0, 0, 0.85);
   justify-content: center;
-}
-.japper:hover {
-  color: #1b4f79;
-}
-/*style cua chu o cac button cua menu*/
-.lesson {
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 300;
-  font-size: 30px;
-  line-height: 36px;
-  display: flex;
-  align-items: center;
-  text-align: center;
   color: #323030;
-  justify-content: center;
 }
 
+/*font*/
+@font-face {
+  font-family: 'MyCustomFont';
+  src: url('../public/avenir_ff/AvenirLTStd-Black.otf') format('truetype');
+
+  font-family: 'MyCustomFont1';
+  src: url('../public/avenir_ff/AvenirLTStd-Roman.otf') format('truetype');
+}
